+++ apps/rc-app/src/app.ts
@@ -8,12 +8,15 @@ import { emitCustomEvent } from "@pagopa/azure-tracing/logger";
 import { makeApplicationInsightsLogger } from "@pagopa/hexagonal-core/adapters/logger";
 import fastify from "fastify";
 import { type RedisClientType, createClient } from "redis";
+import { ulid } from "ulid";
 
 import { AppConfig } from "./adapters/inbound/config/config.js";
+import { mountCreateRcConfigurationHandler } from "./adapters/inbound/fastify/create-rc-configuration.handler.js";
 import { mountGetPublicRcConfigurationHandler } from "./adapters/inbound/fastify/get-public-rc-configuration.handler.js";
 import { mountGetRcConfigurationHandler } from "./adapters/inbound/fastify/get-rc-configuration.handler.js";
 import { mountHealthcheckHandler } from "./adapters/inbound/fastify/healthcheck.handler.js";
 import { mountInfoHandler } from "./adapters/inbound/fastify/info.handler.js";
+import { mountListRcConfigurationHandler } from "./adapters/inbound/fastify/list-rc-configuration.handler.js";
 import { mountUpdateRcConfigurationHandler } from "./adapters/inbound/fastify/update-rc-configuration.handler.js";
 import { CosmosClientHealthcheckAdapter } from "./adapters/outbound/healthcheckers/cosmos.adapter.js";
 import { LoggerHealthcheckAdapter } from "./adapters/outbound/healthcheckers/logger.adapter.js";
@@ -22,10 +25,13 @@ import { PackageJsonAppInfoReader } from "./adapters/outbound/package-json/packa
 import { RCConfigurationCosmosAdapter } from "./adapters/outbound/rc-configurations/rc-configuration.adapter.js";
 import { RCConfigurationCacheAdapter } from "./adapters/outbound/rc-configurations/rc-configuration-cache.adapter.js";
 import { CachingRemoteContentRepository } from "./adapters/outbound/rc-configurations/rc-configuration-caching.adapter.js";
+import { UserRCConfigurationCosmosAdapter } from "./adapters/outbound/rc-configurations/user-rc-configuration.adapter.js";
+import { makeCreateRcConfigurationUseCase } from "./application/use-cases/create-rc-configuration.use-case.js";
 import { makeGetPublicRcConfigurationUseCase } from "./application/use-cases/get-public-rc-configuration.use-case.js";
 import { makeGetRcConfigurationUseCase } from "./application/use-cases/get-rc-configuration.use-case.js";
 import { makeHealthcheckUseCase } from "./application/use-cases/healthcheck.use-case.js";
 import { makeGetInfoUseCase } from "./application/use-cases/info.use-case.js";
+import { makeListRcConfigurationUseCase } from "./application/use-cases/list-rc-confguration.use-case.js";
 import { makeUpdateRcConfigurationUseCase } from "./application/use-cases/update-rc-configuration.use-case.js";
 
 export const createApp = async (
@@ -96,6 +102,15 @@ export const createApp = async (
     },
   });
 
+  const remoteContentRepository = new CachingRemoteContentRepository(
+    new RCConfigurationCosmosAdapter(
+      commonCosmosClient,
+      config.REMOTE_CONTENT_COSMOS_DATABASE_NAME,
+    ),
+    new RCConfigurationCacheAdapter(redisClient, logger),
+    config.RC_CONFIGURATION_CACHE_TTL,
+  );
+
   redisClient.on("error", (err) => {
     server.log.error({ err }, "redis error");
   });
@@ -130,9 +145,19 @@ export const createApp = async (
     config.INTERNAL_USER_ID,
   );
 
-  mountUpdateRcConfigurationHandler(
+  mountCreateRcConfigurationHandler(
     server,
-    makeUpdateRcConfigurationUseCase(
+    makeCreateRcConfigurationUseCase(remoteContentRepository, ulid),
+    config.INTERNAL_USER_ID,
+  );
+
+  mountListRcConfigurationHandler(
+    server,
+    makeListRcConfigurationUseCase(
+      new UserRCConfigurationCosmosAdapter(
+        commonCosmosClient,
+        config.REMOTE_CONTENT_COSMOS_DATABASE_NAME,
+      ),
       new CachingRemoteContentRepository(
         new RCConfigurationCosmosAdapter(
           commonCosmosClient,
@@ -141,8 +166,13 @@ export const createApp = async (
         new RCConfigurationCacheAdapter(redisClient, logger),
         config.RC_CONFIGURATION_CACHE_TTL,
       ),
-      logger,
     ),
+    config.INTERNAL_USER_ID,
+  );
+
+  mountUpdateRcConfigurationHandler(
+    server,
+    makeUpdateRcConfigurationUseCase(remoteContentRepository, logger),
     config.INTERNAL_USER_ID,
   );